+++ lib/jss/api_object/management_history/audit_event.rb
@@ -1,0 +1,45 @@
+#
+module JSS
+
+  #
+  module ManagementHistory
+
+    # AuditEvent - an auditiable event in a Jamf Object's Management History
+    #
+    # This should only be instantiated by the ManagementHistory.audits method
+    # when mixed in to Computers or Mobile devices.
+    #
+    # That method will return an array of these objects.
+    #
+    class AuditEvent < ImmutableStruct.new(
+
+      :event,
+      :username,
+      :date_time_epoch
+    )
+
+      # @!attribute [r] event
+      #  @return [String] The description of the event.
+
+      alias description event
+
+      # @!attribute [r] username
+      #  @return [String] The name of the JSS user who caused the event
+
+      alias admin username
+
+      # @!attribute [r] date_time_epoch
+      #  @return [Integer] When the event occured on the server, as
+      #    a unix epoch timestamp with milliseconds
+
+      #  @return [Time] When the event occured on the server, as a ruby Time object
+      #
+      def date_time
+        JSS.epoch_to_time @date_time_epoch if @date_time_epoch
+      end
+
+    end # AuditEvent
+
+  end #   module ManagementHistory
+
+end # module JSS